Merge variables if the scope of one is contained within another
This improves the output significantly - by merging in more cases, we
avoid the need to use suffixes (e.g. srcOff2, srcOff3, etc. in the
ArrayUtils class) for what is really the same variable.

Restricting the merging to scopes contained in another prevents variable
declarations for two distinct variables sharing a local variable index
from being lifted outside of an if/else block. This often prevents us
from being able to combine a series of if/else blocks into a longer
change, thus increasing the indentation level.